move array/graph sources to array/graph stores
