--- dacapo/experiments/datasplit/datasets/array_source_config.py
@@ -1,10 +1,0 @@
-import attr
-
-
-@attr.s
-class ArraySourceConfig:
-    """Base class for array source configurations. Each subclass of an
-    `ArraySource` should have a corresponding config class derived from
-    `ArraySourceConfig`.
-    """
-    pass
