--- dacapo/experiments/datasplit/datasets/graph_source_config.py
@@ -1,10 +1,0 @@
-import attr
-
-
-@attr.s
-class GraphSourceConfig:
-    """Base class for graph source configurations. Each subclass of a
-    `GraphSource` should have a corresponding config class derived from
-    `GraphSourceConfig`.
-    """
-    pass
