+++ dacapo/experiments/datasplit/datasets/graphstores/graph_source_config.py
@@ -1,0 +1,10 @@
+import attr
+
+
+@attr.s
+class GraphStoreConfig:
+    """Base class for graph store configurations. Each subclass of a
+    `GraphStore` should have a corresponding config class derived from
+    `GraphStoreConfig`.
+    """
+    pass
